feat: move config parsing from apis to internal/otelconfig
