+++ apis/v1alpha1/convert.go
@@ -13,6 +13,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
+	"github.com/open-telemetry/opentelemetry-operator/internal/otelconfig"
 )
 
 var _ conversion.Convertible = &OpenTelemetryCollector{}
@@ -301,7 +302,7 @@ func tov1alpha1Ports(in []v1beta1.PortsSpec) []PortsSpec {
 
 func tov1alpha1(in v1beta1.OpenTelemetryCollector) (*OpenTelemetryCollector, error) {
 	c := in.DeepCopy()
-	configYaml, err := c.Spec.Config.Yaml()
+	configYaml, err := otelconfig.Yaml(&c.Spec.Config)
 	if err != nil {
 		return nil, err
 	}
