+++ internal/manifests/collector/container.go
@@ -18,6 +18,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/internal/autodetect/certmanager"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 	"github.com/open-telemetry/opentelemetry-operator/internal/naming"
+	"github.com/open-telemetry/opentelemetry-operator/internal/otelconfig"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/constants"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/featuregate"
 )
@@ -93,19 +94,19 @@ func Container(cfg config.Config, logger logr.Logger, otelcol v1beta1.OpenTeleme
 		}
 	}
 
-	livenessProbe, livenessProbeErr := otelcol.Spec.Config.GetLivenessProbe(logger)
+	livenessProbe, livenessProbeErr := otelconfig.GetLivenessProbe(&otelcol.Spec.Config, logger)
 	if livenessProbeErr != nil {
 		logger.Error(livenessProbeErr, "cannot create liveness probe.")
 	} else {
 		defaultProbeSettings(livenessProbe, otelcol.Spec.LivenessProbe)
 	}
-	readinessProbe, readinessProbeErr := otelcol.Spec.Config.GetReadinessProbe(logger)
+	readinessProbe, readinessProbeErr := otelconfig.GetReadinessProbe(&otelcol.Spec.Config, logger)
 	if readinessProbeErr != nil {
 		logger.Error(readinessProbeErr, "cannot create readiness probe.")
 	} else {
 		defaultProbeSettings(readinessProbe, otelcol.Spec.ReadinessProbe)
 	}
-	startupProbe, startupProbeErr := otelcol.Spec.Config.GetStartupProbe(logger)
+	startupProbe, startupProbeErr := otelconfig.GetStartupProbe(&otelcol.Spec.Config, logger)
 	if startupProbeErr != nil {
 		logger.Error(startupProbeErr, "cannot create startup probe.")
 	} else {
@@ -132,7 +133,7 @@ func Container(cfg config.Config, logger logr.Logger, otelcol v1beta1.OpenTeleme
 
 func getConfigContainerPorts(logger logr.Logger, conf v1beta1.Config) ([]corev1.ContainerPort, error) {
 	ports := []corev1.ContainerPort{}
-	ps, err := conf.GetAllPorts(logger)
+	ps, err := otelconfig.GetAllPorts(&conf, logger)
 	if err != nil {
 		return ports, err
 	}
@@ -158,7 +159,7 @@ func getConfigContainerPorts(logger logr.Logger, conf v1beta1.Config) ([]corev1.
 		}
 	}
 
-	_, metricsPort, err := conf.Service.MetricsEndpoint(logger)
+	_, metricsPort, err := otelconfig.MetricsEndpoint(&conf.Service, logger)
 	if err != nil {
 		logger.Info("couldn't determine metrics port from configuration, using 8888 default value", "error", err)
 		metricsPort = 8888
@@ -357,7 +358,7 @@ func getInferredContainerEnvVars(otelcol v1beta1.OpenTelemetryCollector, logger 
 		)
 	}
 
-	if configEnvVars, err := otelcol.Spec.Config.GetEnvironmentVariables(logger); err != nil {
+	if configEnvVars, err := otelconfig.GetEnvironmentVariables(&otelcol.Spec.Config, logger); err != nil {
 		logger.Error(err, "could not get the environment variables from the config")
 	} else {
 		envVars = append(envVars, configEnvVars...)
