+++ internal/webhook/collector_webhook.go
@@ -23,6 +23,7 @@ import (
 	ta "github.com/open-telemetry/opentelemetry-operator/internal/manifests/targetallocator/adapters"
 	"github.com/open-telemetry/opentelemetry-operator/internal/metrics"
 	"github.com/open-telemetry/opentelemetry-operator/internal/naming"
+	"github.com/open-telemetry/opentelemetry-operator/internal/otelconfig"
 	"github.com/open-telemetry/opentelemetry-operator/internal/rbac"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/featuregate"
 )
@@ -104,7 +105,7 @@ func (c CollectorWebhook) Default(_ context.Context, otelcol *v1beta1.OpenTeleme
 	// TLS defaults are applied at reconciliation time (ConfigMap generation) so that
 	// existing collectors automatically get updated TLS settings when the operator
 	// restarts after a cluster TLS profile change.
-	events, err := otelcol.Spec.Config.ApplyDefaults(c.logger)
+	events, err := otelconfig.ApplyDefaults(&otelcol.Spec.Config, c.logger)
 	if err != nil {
 		return err
 	}
@@ -167,7 +168,7 @@ func (c CollectorWebhook) ValidateDelete(ctx context.Context, otelcol *v1beta1.O
 func (c CollectorWebhook) Validate(ctx context.Context, r *v1beta1.OpenTelemetryCollector) (admission.Warnings, error) {
 	warnings := admission.Warnings{}
 
-	nullObjects := r.Spec.Config.NullObjects()
+	nullObjects := otelconfig.NullObjects(&r.Spec.Config)
 	if len(nullObjects) > 0 {
 		warnings = append(warnings, fmt.Sprintf("Collector config spec.config has null objects: %s. For compatibility with other tooling, such as kustomize and kubectl edit, it is recommended to use empty objects e.g. batch: {}.", strings.Join(nullObjects, ", ")))
 	}
@@ -220,7 +221,7 @@ func (c CollectorWebhook) Validate(ctx context.Context, r *v1beta1.OpenTelemetry
 	if err := ValidatePorts(r.Spec.Ports); err != nil {
 		return warnings, err
 	}
-	ports, errPorts := r.Spec.Config.GetAllPorts(c.logger)
+	ports, errPorts := otelconfig.GetAllPorts(&r.Spec.Config, c.logger)
 	if errPorts != nil {
 		return warnings, fmt.Errorf("the OpenTelemetry config is incorrect. The port numbers are invalid: %w", errPorts)
 	}
@@ -295,7 +296,7 @@ func (c CollectorWebhook) Validate(ctx context.Context, r *v1beta1.OpenTelemetry
 	}
 
 	if c.fips != nil {
-		components := r.Spec.Config.GetEnabledComponents()
+		components := otelconfig.GetEnabledComponents(&r.Spec.Config)
 		if notAllowedComponents := c.fips.DisabledComponents(components[v1beta1.KindReceiver], components[v1beta1.KindExporter], components[v1beta1.KindProcessor], components[v1beta1.KindExtension]); notAllowedComponents != nil {
 			return nil, fmt.Errorf("the collector configuration contains not FIPS compliant components: %s. Please remove it from the config", notAllowedComponents)
 		}
@@ -317,7 +318,7 @@ func (c CollectorWebhook) validateTargetAllocatorConfig(ctx context.Context, r *
 		return nil, fmt.Errorf("target allocation strategy %s is only supported in OpenTelemetry Collector mode %s", v1beta1.TargetAllocatorAllocationStrategyPerNode, v1beta1.ModeDaemonSet)
 	}
 
-	cfgYaml, err := r.Spec.Config.Yaml()
+	cfgYaml, err := otelconfig.Yaml(&r.Spec.Config)
 	if err != nil {
 		return nil, err
 	}
