+++ pkg/collector/upgrade/v0_122_0.go
@@ -5,16 +5,17 @@ package upgrade
 
 import (
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
+	"github.com/open-telemetry/opentelemetry-operator/internal/otelconfig"
 )
 
 func upgrade0_122_0(u VersionUpgrade, otelcol *v1beta1.OpenTelemetryCollector) (*v1beta1.OpenTelemetryCollector, error) {
-	tel := otelcol.Spec.Config.Service.GetTelemetry(u.Log)
+	tel := otelconfig.GetTelemetry(&otelcol.Spec.Config.Service, u.Log)
 
 	if tel == nil || tel.Metrics.Address == "" {
 		return otelcol, nil
 	}
 
-	host, port, err := otelcol.Spec.Config.Service.MetricsEndpoint(u.Log)
+	host, port, err := otelconfig.MetricsEndpoint(&otelcol.Spec.Config.Service, u.Log)
 	if err != nil {
 		return otelcol, err
 	}
@@ -26,10 +27,10 @@ func upgrade0_122_0(u VersionUpgrade, otelcol *v1beta1.OpenTelemetryCollector) (
 	// differently from explicitly empty ones. By assigning "", we ensure the configuration
 	// is updated correctly when the resource is persisted.
 	tel.Metrics.Address = ""
-	reader := v1beta1.AddPrometheusMetricsEndpoint(host, port)
+	reader := otelconfig.AddPrometheusMetricsEndpoint(host, port)
 	tel.Metrics.Readers = append(tel.Metrics.Readers, reader)
 
-	otelcol.Spec.Config.Service.Telemetry, err = tel.ToAnyConfig()
+	otelcol.Spec.Config.Service.Telemetry, err = otelconfig.TelemetryToAnyConfig(tel)
 	if err != nil {
 		return otelcol, err
 	}
