+++ pkg/sidecar/pod_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1beta1"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 	"github.com/open-telemetry/opentelemetry-operator/internal/naming"
+	"github.com/open-telemetry/opentelemetry-operator/internal/otelconfig"
 )
 
 var logger = logf.Log.WithName("unit-tests")
@@ -59,7 +60,7 @@ func TestAddNativeSidecar(t *testing.T) {
 		},
 	}
 
-	otelcolYaml, err := otelcol.Spec.Config.Yaml()
+	otelcolYaml, err := otelconfig.Yaml(&otelcol.Spec.Config)
 	require.NoError(t, err)
 	cfg := config.Config{
 		CollectorImage: "some-default-image",
@@ -188,7 +189,7 @@ func TestAddSidecarWhenNoSidecarExists(t *testing.T) {
 		},
 	}
 
-	otelcolYaml, err := otelcol.Spec.Config.Yaml()
+	otelcolYaml, err := otelconfig.Yaml(&otelcol.Spec.Config)
 	require.NoError(t, err)
 	cfg := config.Config{
 		CollectorImage: "some-default-image",
